llava/requirements.txt
transformers
flask
Pillow
torch
torchvision
opencv-python
matplotlib
numpy
scipy
